Convert tags to resource identifiers using a pure function.
`ImmutableMap`s are sometimes used for `Batch.Point`s.
When attempting to convert known tags to resource identifiers with the
old implementation, this would cause an exception trying to mutate the `ImmutableMap`.

Using a pure function for this is more verbose, but its so much safer not having to
worry about mutating state.