Add an expiring writeCache
An ExpiringCache is useful to reduce the number of metrics being published to a topic. Typically processing and de-deduplication of the "metadata" around a metric only needs to happen every so often. In the case of using Heroic, this "metadata" only needs to be published at the interval of a new index being created.