Properly shade gRPC modules when packaging
The maven-shade-plugin runs as part of the agent packaging. With the
new version and the ServicesResourceTransformer, it will merge the
service files. This shading issue only caused failures when the
modules are packaged, since running through an IDE never triggers the
shading. Closes #237

See https://github.com/grpc/grpc-java/issues/3141 for discussion of
others running into this issue for gRPC.

The exclusion removals are not necessary for this fix, just additional
cleanup while I was investigating.